move all generators over to particle-cli and remove yo referencing in preference to local file running of yeoman generators
