+++ packages/generator-particle-cli/package.json
@@ -1,10 +1,15 @@
 {
-  "name": "@phase2/generator-particle-cli",
+  "name": "@phase2/particle-cli",
   "version": "0.0.1",
   "description": "Generates the file structure for particle",
   "files": [
-    "lib/generators"
+    "lib/src/generators/app",
+    "lib/src/generators/storybook"
   ],
+  "bin": {
+    "@phase2/particle-cli": "lib/src/bin/particle-cli.js",
+    "particle-cli": "lib/src/bin/particle-cli.js"
+  },
   "keywords": [
     "yeoman-generator",
     "particle"
@@ -13,19 +18,21 @@
   "repository": {
     "type": "git",
     "url": "git+https://github.com/phase2/particle.git",
-    "directory": "packages/generator-particle"
+    "directory": "packages/particle-cli"
   },
   "dependencies": {
     "@phase2/particle-types": "file:../types",
     "chalk": "^4.1.0",
+    "commander": "^5.1.0",
+    "inquirer": "^7.2.0",
     "lodash.merge": "^4.6.2",
-    "yeoman-generator": "^4.11.0"
+    "yeoman-generator": "^4.11.0",
+    "yo": "^3.1.1"
   },
   "devDependencies": {
     "@types/lodash.merge": "^4.6.6",
     "@types/yeoman-generator": "^3.1.4",
-    "typescript": "^3.9.6",
-    "yo": "^3.1.1"
+    "typescript": "^3.9.6"
   },
   "scripts": {
     "tsc": "tsc",
