--- packages/generator-particle-cli/test/generators/app/index.d.ts
@@ -1,1 +1,0 @@
-export {};
