--- packages/generator-particle-cli/test/generators/storybook/index.d.ts
@@ -1,1 +1,0 @@
-export {};
