+++ packages/generator-particle-cli/tsconfig.json
@@ -2,7 +2,7 @@
   "extends": "../../tsconfig.json",
   "compilerOptions": {
     "outDir": "lib",
-    "rootDir": "src"
+    "rootDir": "./"
   },
-  "include": ["./src/generators"]
+  "include": ["./src/generators", "./src/bin", "./package.json"]
 }
