--- packages/particle-cli/package.json
@@ -1,23 +1,0 @@
-{
-  "name": "new-project",
-  "version": "1.0.0",
-  "main": "index.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "build:storybook": "build-storybook -c ./apps/storybook",
-    "dev:storybook": "start-storybook -p 6006 -c ./apps/storybook"
-  },
-  "keywords": [],
-  "author": "",
-  "license": "ISC",
-  "description": "",
-  "dependencies": {},
-  "devDependencies": {
-    "@storybook/react": "^5.3.19",
-    "@storybook/addon-viewport": "^5.3.19",
-    "@storybook/addon-knobs": "^5.3.19",
-    "@storybook/addon-actions": "^5.3.19",
-    "@storybook/addon-a11y": "^5.3.19",
-    "@storybook/addon-links": "^5.3.19"
-  }
-}
