--- packages/particle-cli/tsconfig.json
@@ -1,8 +1,0 @@
-{
-  "extends": "../../tsconfig.json",
-  "compilerOptions": {
-    "outDir": "lib",
-    "rootDir": "./"
-  },
-  "include": ["./src", "./bin", "./package.json"]
-}
